fix: code restructure
