+++ src/main/automationExecutor.ts
@@ -1,146 +1,90 @@
 import { AutomationStep, StepAIAnthropic, StepAIOllama, StepAIOpenAI } from "@app-types/steps";
-import axios from "axios";
-import crypto from "crypto";
+import { debugLogger } from "@main/debugLogger";
+import {
+  AiStepHandler,
+  ApiCallHandler,
+  BrowserStepHandler,
+  ConditionalEvaluator,
+  DiscordStepHandler,
+  TwitterStepHandler,
+  VariableManager,
+  VariableOperationHandler,
+} from "@main/handlers";
+import { HeadlessExecutor } from "@main/headlessExecutor";
 import { BrowserWindow } from "electron";
-import fs from "fs";
-import { getCredential } from "./credentialsStore";
-import { debugLogger } from "./debugLogger";
-import { HeadlessExecutor } from "./headlessExecutor";
 
 /**
  * Handles execution of automation steps in the browser window
+ * Delegates step execution to specialized handlers for cleaner code organization
  */
 export class AutomationExecutor {
-  /** Unified variable storage - automatically typed by input
-   * Variables are automatically typed based on input:
-   * "42" → number, "true" → boolean, objects/arrays stay as-is, else string
-   */
-  private variables: Record<string, unknown> = {};
+  private variableManager: VariableManager;
+  private browserHandler: BrowserStepHandler;
+  private apiCallHandler: ApiCallHandler;
+  private aiHandler: AiStepHandler;
+  private twitterHandler: TwitterStepHandler;
+  private discordHandler: DiscordStepHandler;
+  private variableOperationHandler: VariableOperationHandler;
+  private conditionalEvaluator: ConditionalEvaluator;
+
+  constructor() {
+    this.variableManager = new VariableManager();
+    this.browserHandler = new BrowserStepHandler();
+    this.apiCallHandler = new ApiCallHandler();
+    this.aiHandler = new AiStepHandler();
+    this.twitterHandler = new TwitterStepHandler();
+    this.discordHandler = new DiscordStepHandler();
+    this.variableOperationHandler = new VariableOperationHandler();
+    this.conditionalEvaluator = new ConditionalEvaluator();
+  }
 
   /**
    * Initialize executor variable context
    */
   initVariables(vars?: Record<string, unknown>) {
-    this.variables = { ...(vars || {}) };
+    this.variableManager.initVariables(vars);
   }
 
   /**
    * Expose a shallow copy of current variables (for logging / IPC)
    */
   getVariables(): Record<string, unknown> {
-    return { ...this.variables };
+    return this.variableManager.getVariables();
   }
 
   /**
    * Get a variable value, supporting dot notation and array indexing
-   * e.g., "userinfo.name" → nested property
-   * e.g., "users[0]" → first array element
-   * e.g., "users[0].name" → property of array element
    */
   private getVariableValue(path: string): unknown {
-    // Parse path into tokens: handles both . and [index] syntax
-    const tokens: (string | number)[] = [];
-    let current = "";
-    let i = 0;
-
-    while (i < path.length) {
-      const char = path[i];
-
-      if (char === ".") {
-        if (current) tokens.push(current);
-        current = "";
-        i++;
-      } else if (char === "[") {
-        // Handle array index [0], [1], etc.
-        if (current) tokens.push(current);
-        current = "";
-        i++;
-        let indexStr = "";
-        while (i < path.length && path[i] !== "]") {
-          indexStr += path[i];
-          i++;
-        }
-        if (path[i] === "]") i++; // Skip closing ]
-        const index = parseInt(indexStr, 10);
-        if (!isNaN(index)) tokens.push(index);
-      } else {
-        current += char;
-        i++;
-      }
-    }
-
-    if (current) tokens.push(current);
-
-    // Navigate through tokens
-    if (tokens.length === 0) return "";
-
-    let value: unknown = this.variables[String(tokens[0])];
-
-    for (let i = 1; i < tokens.length; i++) {
-      if (value === null || value === undefined) return "";
-
-      const token = tokens[i];
-
-      if (typeof token === "number") {
-        // Array indexing
-        if (Array.isArray(value)) {
-          value = value[token];
-        } else {
-          return "";
-        }
-      } else {
-        // Property access
-        if (typeof value === "object") {
-          value = (value as Record<string, unknown>)[token];
-        } else {
-          return "";
-        }
-      }
-    }
-
-    return value;
+    return this.variableManager.getVariableValue(path);
   }
 
   /**
    * Auto-detect and convert variable value to appropriate type
    */
   private parseValue(input: string): unknown {
-    // Try JSON parse first (objects, arrays, null)
-    try {
-      const parsed = JSON.parse(input);
-      return parsed;
-    } catch {
-      // Not JSON, try other types
-    }
-
-    // Boolean conversion
-    if (input === "true") return true;
-    if (input === "false") return false;
-
-    // Number conversion
-    if (!isNaN(Number(input)) && input !== "") {
-      return Number(input);
-    }
-
-    // Default to string
-    return input;
+    return this.variableManager.parseValue(input);
   }
 
   /**
    * Replace {{varName}} and {{varName.property}} tokens in a string with current variable values.
-   * Supports dot notation for nested property access and array indexing.
-   * Variables are automatically typed based on their value.
-   * Examples: {{var}}, {{var.prop}}, {{array[0]}}, {{array[0].name}}
    */
   private substituteVariables(input?: string): string {
-    if (!input) return "";
-    // Match {{varName}}, {{varName.property}}, {{array[0]}}, etc.
-    return input.replace(/\{\{\s*([a-zA-Z0-9_[\].]+)\s*\}\}/g, (_, path) => {
-      const value = this.getVariableValue(path);
-      if (value === null || value === undefined) return "";
-      if (typeof value === "object") return JSON.stringify(value);
-      return String(value);
-    });
+    return this.variableManager.substituteVariables(input);
+  }
+
+  /**
+   * Set a variable value
+   */
+  private setVariable(key: string, value: unknown) {
+    this.variableManager.setVariable(key, value);
+  }
+
+  /**
+   * Get the underlying variables object for passes to handlers
+   */
+  private getVariablesObject(): Record<string, unknown> {
+    return this.variableManager.getVariables();
   }
 
   /**
@@ -154,7 +98,6 @@ export class AutomationExecutor {
     headlessExecutor: HeadlessExecutor | null | undefined,
     step: AutomationStep
   ): Promise<unknown> {
-    const wc = browserWindow?.webContents;
     const startTime = performance.now();
 
     debugLogger.debug("Step Execution", `Starting ${step.type} step`, {
@@ -164,907 +107,300 @@ export class AutomationExecutor {
 
     try {
       let result: unknown;
+      const variables = this.getVariablesObject();
 
       switch (step.type) {
-        case "navigate": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            const url = this.substituteVariables(step.value);
-            if (!wc) throw new Error("Browser window required for navigate step");
-            debugLogger.debug("Navigate", `Loading URL: ${url}`);
-            await wc.loadURL(url);
-            result = undefined;
-          }
-          break;
-        }
-
-        case "click": {
-          const clickSelector = this.substituteVariables(step.selector);
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for click step");
-            debugLogger.debug("Click", `Clicking element with selector: ${clickSelector}`);
-            await wc.executeJavaScript(`
-            (() => {
-              const sel = ${JSON.stringify(clickSelector)};
-              let el = null;
-              try { el = document.querySelector(sel); } catch(e) {}
-              if (!el) {
-                try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-              }
-              if (el) el.click();
-            })();
-          `);
-            result = undefined;
-          }
-          break;
-        }
-
-        case "type": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for type step");
-            const typeSelector = this.substituteVariables(step.selector);
-            const typeValue = this.substituteVariables(step.value);
-            debugLogger.debug("Type", `Typing into selector: ${typeSelector}`, {
-              value: typeValue,
-            });
-            await wc.executeJavaScript(`
-            (() => {
-              const sel = ${JSON.stringify(typeSelector)};
-              let el = null;
-              try { el = document.querySelector(sel); } catch(e) {}
-              if (!el) {
-                try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-              }
-              if (el) {
-                el.focus();
-                el.value = ${JSON.stringify(typeValue)};
-                el.dispatchEvent(new Event('input', { bubbles: true }));
-              }
-            })();
-          `);
-            result = undefined;
-          }
-          break;
-        }
-
-        case "wait": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            const waitStr = this.substituteVariables(step.value);
-            const ms = parseInt(waitStr) * 1000;
-            debugLogger.debug("Wait", `Waiting for ${waitStr} seconds`);
-            await new Promise((resolve) => {
-              setTimeout(resolve, isNaN(ms) ? 0 : ms);
-            });
-            result = undefined;
-          }
-          break;
-        }
-
-        case "screenshot": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for screenshot step");
-            debugLogger.debug("Screenshot", "Capturing page screenshot");
-            const img = await wc.capturePage();
-            const timestamp = new Date().toISOString().replace(/[-:.]/g, "").slice(0, 15);
-            const savePath = step.savePath || `screenshot_${timestamp}.png`;
-            await fs.promises.writeFile(savePath, img.toPNG());
-            debugLogger.info("Screenshot", `Screenshot saved to: ${savePath}`);
-            result = img.toPNG().toString("base64");
-          }
-          break;
-        }
-
-        case "extract": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(
-              step,
-              this.substituteVariables.bind(this),
-              this.variables
-            );
-          } else {
-            if (!wc) throw new Error("Browser window required for extract step");
-            const extractSelector = this.substituteVariables(step.selector);
-            debugLogger.debug("Extract", `Extracting text from selector: ${extractSelector}`);
-            const extracted = await wc.executeJavaScript(`
-            (() => {
-              const sel = ${JSON.stringify(extractSelector)};
-              let el = null;
-              try { el = document.querySelector(sel); } catch(e) {}
-              if (!el) {
-                try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-              }
-              return el?.innerText || "";
-            })();
-          `);
-            if (step.storeKey) {
-              this.variables[step.storeKey] = extracted;
-              debugLogger.debug("Extract", `Stored extracted value in variable: ${step.storeKey}`, {
-                value: extracted,
-              });
-            }
-            result = extracted;
-          }
-          break;
-        }
-
-        case "apiCall": {
-          try {
-            const url = this.substituteVariables(step.url);
-            const method = step.method || "GET";
-            debugLogger.debug("API Call", `Making ${method} request to: ${url}`);
-
-            const rawBody = step.body ? this.substituteVariables(step.body) : undefined;
-            let data: unknown = undefined;
-            if (rawBody) {
-              try {
-                data = JSON.parse(rawBody);
-              } catch {
-                data = rawBody;
-              }
-            }
-            const headers: Record<string, string> = {};
-            if (step.headers) {
-              for (const [k, v] of Object.entries(step.headers as Record<string, unknown>)) {
-                headers[k] = this.substituteVariables(String(v));
-              }
-            }
-
-            debugLogger.debug("API Call", `Request headers and data`, { headers, hasBody: !!data });
-            const response = await axios({ method, url, data, headers });
-            const dataOut = response.data;
-
-            debugLogger.debug("API Call", `Response received`, {
-              status: response.status,
-              dataLength: JSON.stringify(dataOut).length,
-            });
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = dataOut;
-              debugLogger.debug("API Call", `Stored API response in variable: ${step.storeKey}`);
-            }
-            result = dataOut;
-          } catch (error) {
-            debugLogger.error("API Call", "API call failed", error);
-            throw error;
-          }
-          break;
-        }
-
-        case "aiOpenAI": {
-          const aiResult = await this.executeAiOpenAI(step);
-          if (step.storeKey) {
-            this.variables[step.storeKey] = aiResult;
-            debugLogger.debug("AI OpenAI", `Stored response in variable: ${step.storeKey}`);
-          }
-          result = aiResult;
-          break;
-        }
-
-        case "aiAnthropic": {
-          const aiResult = await this.executeAiAnthropic(step);
-          if (step.storeKey) {
-            this.variables[step.storeKey] = aiResult;
-            debugLogger.debug("AI Anthropic", `Stored response in variable: ${step.storeKey}`);
-          }
-          result = aiResult;
-          break;
-        }
-
-        case "aiOllama": {
-          const aiResult = await this.executeAiOllama(step);
-          if (step.storeKey) {
-            this.variables[step.storeKey] = aiResult;
-            debugLogger.debug("AI Ollama", `Stored response in variable: ${step.storeKey}`);
-          }
-          result = aiResult;
-          break;
-        }
-
-        case "discordSendMessage": {
-          const channelId = this.substituteVariables(step.channelId);
-          const content = this.substituteVariables(step.content);
-          const { botToken } = await this.resolveDiscordCredentials(step);
-
-          if (!botToken) {
-            throw new Error("Discord bot token is required to send messages");
-          }
-
-          const payload: Record<string, unknown> = { content, tts: !!step.tts };
-
-          const response = await this.discordRequest(
-            "POST",
-            `/channels/${channelId}/messages`,
-            payload,
-            botToken
+        // Browser interaction steps
+        case "navigate":
+          result = await this.browserHandler.executeNavigate(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
           );
-
-          if (step.storeKey) {
-            this.variables[step.storeKey] = response.data;
-          }
-
-          result = response.data;
           break;
-        }
 
-        case "discordSendWebhook": {
-          const content = this.substituteVariables(step.content);
-          const resolvedWebhook = this.substituteVariables(step.webhookUrl || "");
-
-          if (!resolvedWebhook) {
-            throw new Error("Webhook URL is required for Discord webhook step");
-          }
-
-          const payload: Record<string, unknown> = {
-            content,
-            username: step.username ? this.substituteVariables(step.username) : undefined,
-            avatar_url: step.avatarUrl ? this.substituteVariables(step.avatarUrl) : undefined,
-            tts: !!step.tts,
-          };
-
-          if (step.embedsJson) {
-            try {
-              payload.embeds = JSON.parse(this.substituteVariables(step.embedsJson));
-            } catch (error) {
-              debugLogger.error("Discord Send Webhook", "Invalid embeds JSON", error);
-              throw new Error("Invalid embeds JSON for Discord webhook");
-            }
-          }
-
-          const response = await axios.post(resolvedWebhook, payload, {
-            headers: { "Content-Type": "application/json" },
-          });
-
-          if (step.storeKey) {
-            this.variables[step.storeKey] = response.data;
-          }
-
-          result = response.data;
-          break;
-        }
-
-        case "discordReactMessage": {
-          const channelId = this.substituteVariables(step.channelId);
-          const messageId = this.substituteVariables(step.messageId);
-          const emoji = encodeURIComponent(this.substituteVariables(step.emoji));
-          const { botToken } = await this.resolveDiscordCredentials(step);
-
-          if (!botToken) {
-            throw new Error("Discord bot token is required to react to messages");
-          }
-
-          await this.discordRequest(
-            "PUT",
-            `/channels/${channelId}/messages/${messageId}/reactions/${emoji}/@me`,
-            undefined,
-            botToken
+        case "click":
+          result = await this.browserHandler.executeClick(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
           );
-
-          result = { success: true };
           break;
-        }
 
-        case "discordGetMessage": {
-          const channelId = this.substituteVariables(step.channelId);
-          const messageId = this.substituteVariables(step.messageId);
-          const { botToken } = await this.resolveDiscordCredentials(step);
-
-          if (!botToken) {
-            throw new Error("Discord bot token is required to fetch messages");
-          }
-
-          const response = await this.discordRequest(
-            "GET",
-            `/channels/${channelId}/messages/${messageId}`,
-            undefined,
-            botToken
+        case "type":
+          result = await this.browserHandler.executeType(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
           );
-
-          if (step.storeKey) {
-            this.variables[step.storeKey] = response.data;
-          }
-
-          result = response.data;
           break;
-        }
 
-        case "discordListMessages": {
-          const channelId = this.substituteVariables(step.channelId);
-          const limit = step.limit && step.limit > 0 ? Math.min(step.limit, 100) : 10;
-          const { botToken } = await this.resolveDiscordCredentials(step);
-
-          if (!botToken) {
-            throw new Error("Discord bot token is required to list messages");
-          }
-
-          const response = await this.discordRequest(
-            "GET",
-            `/channels/${channelId}/messages?limit=${limit}`,
-            undefined,
-            botToken
+        case "wait":
+          result = await this.browserHandler.executeWait(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
           );
-
-          if (step.storeKey) {
-            this.variables[step.storeKey] = response.data;
-          }
-
-          result = response.data;
           break;
-        }
 
-        case "discordDeleteMessage": {
-          const channelId = this.substituteVariables(step.channelId);
-          const messageId = this.substituteVariables(step.messageId);
-          const { botToken } = await this.resolveDiscordCredentials(step);
-
-          if (!botToken) {
-            throw new Error("Discord bot token is required to delete messages");
-          }
-
-          await this.discordRequest(
-            "DELETE",
-            `/channels/${channelId}/messages/${messageId}`,
-            undefined,
-            botToken
+        case "screenshot":
+          result = await this.browserHandler.executeScreenshot(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
           );
-
-          result = { success: true };
           break;
-        }
 
-        case "scroll": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for scroll step");
-            if (step.scrollType === "toElement") {
-              const scrollSelector = this.substituteVariables(step.selector || "");
-              debugLogger.debug("Scroll", `Scrolling to element: ${scrollSelector}`);
-              await wc.executeJavaScript(`
-              (() => {
-                const sel = ${JSON.stringify(scrollSelector)};
-                let el = null;
-                try { el = document.querySelector(sel); } catch(e) {}
-                if (!el) {
-                  try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-                }
-                if (el) el.scrollIntoView({ behavior: "smooth", block: "center" });
-              })();
-            `);
-            } else if (step.scrollType === "byAmount") {
-              debugLogger.debug("Scroll", `Scrolling by ${step.scrollAmount} pixels`);
-              await wc.executeJavaScript(`window.scrollBy(0, ${step.scrollAmount || 0});`);
-            }
-            result = undefined;
+        case "scroll":
+          result = await this.browserHandler.executeScroll(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
+          );
+          break;
+
+        case "fileUpload":
+          result = await this.browserHandler.executeFileUpload(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
+          );
+          break;
+
+        case "hover":
+          result = await this.browserHandler.executeHover(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
+          );
+          break;
+
+        case "selectOption":
+          result = await this.browserHandler.executeSelectOption(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this)
+          );
+          break;
+
+        case "extract":
+          result = await this.browserHandler.executeExtract(
+            browserWindow,
+            headless,
+            headlessExecutor,
+            step,
+            this.substituteVariables.bind(this),
+            variables
+          );
+          break;
+
+        // API Call step
+        case "apiCall":
+          result = await this.apiCallHandler.executeApiCall(
+            step,
+            this.substituteVariables.bind(this),
+            variables
+          );
+          break;
+
+        // AI steps
+        case "aiOpenAI":
+          result = await this.aiHandler.executeAiOpenAI(
+            step as StepAIOpenAI,
+            this.substituteVariables.bind(this)
+          );
+          if ((step as StepAIOpenAI).storeKey) {
+            variables[(step as StepAIOpenAI).storeKey] = result;
+            debugLogger.debug(
+              "AI OpenAI",
+              `Stored response in variable: ${(step as StepAIOpenAI).storeKey}`
+            );
           }
           break;
-        }
 
-        case "selectOption": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for selectOption step");
-            const selectSelector = this.substituteVariables(step.selector);
-            const optionValue = this.substituteVariables(step.optionValue);
-            debugLogger.debug("Select Option", `Selecting option in: ${selectSelector}`, {
-              optionValue,
-            });
-            await wc.executeJavaScript(`
-            (() => {
-              const sel = ${JSON.stringify(selectSelector)};
-              let select = null;
-              try { select = document.querySelector(sel); } catch(e) {}
-              if (!select) {
-                try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); select = r.singleNodeValue; } catch(e2) {}
-              }
-              if (select) {
-                ${step.optionValue ? `select.value = ${JSON.stringify(optionValue)};` : `select.selectedIndex = ${step.optionIndex || 0};`}
-                select.dispatchEvent(new Event('change', { bubbles: true }));
-              }
-            })();
-          `);
-            result = undefined;
+        case "aiAnthropic":
+          result = await this.aiHandler.executeAiAnthropic(
+            step as StepAIAnthropic,
+            this.substituteVariables.bind(this)
+          );
+          if ((step as StepAIAnthropic).storeKey) {
+            variables[(step as StepAIAnthropic).storeKey] = result;
+            debugLogger.debug(
+              "AI Anthropic",
+              `Stored response in variable: ${(step as StepAIAnthropic).storeKey}`
+            );
           }
           break;
-        }
 
-        case "fileUpload": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for fileUpload step");
-            const fuSelector = this.substituteVariables(step.selector);
-            const fuFilePath = this.substituteVariables(step.filePath);
-            debugLogger.debug("File Upload", `Uploading file to: ${fuSelector}`, {
-              filePath: fuFilePath,
-            });
-            await wc.executeJavaScript(`
-            (() => {
-              const sel = ${JSON.stringify(fuSelector)};
-              let input = null;
-              try { input = document.querySelector(sel); } catch(e) {}
-              if (!input) {
-                try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); input = r.singleNodeValue; } catch(e2) {}
-              }
-              if (input) {
-                const file = new File([""], ${JSON.stringify(fuFilePath)}, { type: "text/plain" });
-                const dataTransfer = new DataTransfer();
-                dataTransfer.items.add(file);
-                input.files = dataTransfer.files;
-                input.dispatchEvent(new Event('change', { bubbles: true }));
-              }
-            })();
-          `);
-            result = undefined;
+        case "aiOllama":
+          result = await this.aiHandler.executeAiOllama(
+            step as StepAIOllama,
+            this.substituteVariables.bind(this)
+          );
+          if ((step as StepAIOllama).storeKey) {
+            variables[(step as StepAIOllama).storeKey] = result;
+            debugLogger.debug(
+              "AI Ollama",
+              `Stored response in variable: ${(step as StepAIOllama).storeKey}`
+            );
           }
           break;
-        }
 
-        case "hover": {
-          if (headless) {
-            if (!headlessExecutor) throw new Error("Headless executor not initialized");
-            await headlessExecutor.executeStep(step, this.substituteVariables.bind(this));
-          } else {
-            if (!wc) throw new Error("Browser window required for hover step");
-            const hoverSelector = this.substituteVariables(step.selector);
-            debugLogger.debug("Hover", `Hovering over element: ${hoverSelector}`);
-            await wc.executeJavaScript(`
-            (() => {
-              const sel = ${JSON.stringify(hoverSelector)};
-              let el = null;
-              try { el = document.querySelector(sel); } catch(e) {}
-              if (!el) {
-                try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-              }
-              if (el) {
-                const event = new MouseEvent('mouseover', { bubbles: true });
-                el.dispatchEvent(event);
-              }
-            })();
-          `);
-            result = undefined;
-          }
+        // Discord steps
+        case "discordSendMessage":
+          result = await this.discordHandler.executeSendMessage(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveDiscordCredentials.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "setVariable": {
-          const varName = step.variableName;
-          const rawValue = this.substituteVariables(step.value);
-          const value = this.parseValue(rawValue);
-          this.variables[varName] = value;
-          debugLogger.debug("Set Variable", `Variable '${varName}' set`, { value });
-          result = value;
+        case "discordSendWebhook":
+          result = await this.discordHandler.executeSendWebhook(
+            step,
+            this.substituteVariables.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "modifyVariable": {
-          const name = step.variableName;
-          const raw = this.substituteVariables(step.value);
-          const op = step.operation;
-          const current = this.variables[name];
-
-          if (op === "set") {
-            const value = this.parseValue(raw);
-            this.variables[name] = value;
-            debugLogger.debug("Modify Variable", `Variable '${name}' modified (set)`, { value });
-            result = value;
-          } else if (op === "increment") {
-            const num = typeof current === "number" ? current : parseFloat(String(current) || "0");
-            const by = parseFloat(raw || "1");
-            const res = num + by;
-            this.variables[name] = res;
-            debugLogger.debug("Modify Variable", `Variable '${name}' incremented`, {
-              from: num,
-              by,
-              to: res,
-            });
-            result = res;
-          } else if (op === "decrement") {
-            const num = typeof current === "number" ? current : parseFloat(String(current) || "0");
-            const by = parseFloat(raw || "1");
-            const res = num - by;
-            this.variables[name] = res;
-            debugLogger.debug("Modify Variable", `Variable '${name}' decremented`, {
-              from: num,
-              by,
-              to: res,
-            });
-            result = res;
-          } else if (op === "append") {
-            const res = String(current || "") + raw;
-            this.variables[name] = res;
-            debugLogger.debug("Modify Variable", `Variable '${name}' appended`, { result: res });
-            result = res;
-          }
+        case "discordReactMessage":
+          result = await this.discordHandler.executeReactMessage(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveDiscordCredentials.bind(this)
+          );
           break;
-        }
 
-        case "twitterCreateTweet": {
-          try {
-            const text = this.substituteVariables(step.text);
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Create Tweet", "Creating tweet", { text });
-
-            const body: Record<string, unknown> = { text };
-
-            if (step.replyToTweetId) {
-              body.reply = {
-                in_reply_to_tweet_id: this.substituteVariables(step.replyToTweetId),
-              };
-            }
-
-            if (step.quoteTweetId) {
-              body.quote_tweet_id = this.substituteVariables(step.quoteTweetId);
-            }
-
-            if (step.mediaId) {
-              body.media = { media_ids: [this.substituteVariables(step.mediaId)] };
-            }
-
-            const oauth = this.generateTwitterOAuthHeader(
-              "POST",
-              "https://api.twitter.com/2/tweets",
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.post("https://api.twitter.com/2/tweets", body, {
-              headers: {
-                Authorization: oauth,
-                "Content-Type": "application/json",
-              },
-            });
-
-            debugLogger.debug("Twitter Create Tweet", "Tweet created successfully", {
-              tweetId: response.data.data?.id,
-            });
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data.data;
-            }
-
-            result = response.data.data;
-          } catch (error) {
-            debugLogger.error("Twitter Create Tweet", "Failed to create tweet", error);
-            throw error;
-          }
+        case "discordGetMessage":
+          result = await this.discordHandler.executeGetMessage(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveDiscordCredentials.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "twitterDeleteTweet": {
-          try {
-            const tweetId = this.extractTweetId(this.substituteVariables(step.tweetId));
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Delete Tweet", "Deleting tweet", { tweetId });
-
-            const oauth = this.generateTwitterOAuthHeader(
-              "DELETE",
-              `https://api.twitter.com/2/tweets/${tweetId}`,
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.delete(`https://api.twitter.com/2/tweets/${tweetId}`, {
-              headers: { Authorization: oauth },
-            });
-
-            debugLogger.debug("Twitter Delete Tweet", "Tweet deleted successfully");
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data;
-            }
-
-            result = response.data;
-          } catch (error) {
-            debugLogger.error("Twitter Delete Tweet", "Failed to delete tweet", error);
-            throw error;
-          }
+        case "discordListMessages":
+          result = await this.discordHandler.executeListMessages(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveDiscordCredentials.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "twitterLikeTweet": {
-          try {
-            const tweetId = this.extractTweetId(this.substituteVariables(step.tweetId));
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Like Tweet", "Liking tweet", { tweetId });
-
-            const userOauth = this.generateTwitterOAuthHeader(
-              "GET",
-              "https://api.twitter.com/2/users/me",
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const userResponse = await axios.get("https://api.twitter.com/2/users/me", {
-              headers: { Authorization: userOauth },
-            });
-
-            const userId = userResponse.data.data.id;
-
-            const likeOauth = this.generateTwitterOAuthHeader(
-              "POST",
-              `https://api.twitter.com/2/users/${userId}/likes`,
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.post(
-              `https://api.twitter.com/2/users/${userId}/likes`,
-              { tweet_id: tweetId },
-              {
-                headers: {
-                  Authorization: likeOauth,
-                  "Content-Type": "application/json",
-                },
-              }
-            );
-
-            debugLogger.debug("Twitter Like Tweet", "Tweet liked successfully");
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data;
-            }
-
-            result = response.data;
-          } catch (error) {
-            debugLogger.error("Twitter Like Tweet", "Failed to like tweet", error);
-            throw error;
-          }
+        case "discordDeleteMessage":
+          result = await this.discordHandler.executeDeleteMessage(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveDiscordCredentials.bind(this)
+          );
           break;
-        }
 
-        case "twitterRetweet": {
-          try {
-            const tweetId = this.extractTweetId(this.substituteVariables(step.tweetId));
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Retweet", "Retweeting", { tweetId });
-
-            // Get current user ID
-            const userOauth = this.generateTwitterOAuthHeader(
-              "GET",
-              "https://api.twitter.com/2/users/me",
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const userResponse = await axios.get("https://api.twitter.com/2/users/me", {
-              headers: { Authorization: userOauth },
-            });
-
-            const userId = userResponse.data.data.id;
-
-            const retweetOauth = this.generateTwitterOAuthHeader(
-              "POST",
-              `https://api.twitter.com/2/users/${userId}/retweets`,
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.post(
-              `https://api.twitter.com/2/users/${userId}/retweets`,
-              { tweet_id: tweetId },
-              {
-                headers: {
-                  Authorization: retweetOauth,
-                  "Content-Type": "application/json",
-                },
-              }
-            );
-
-            debugLogger.debug("Twitter Retweet", "Retweeted successfully");
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data;
-            }
-
-            result = response.data;
-          } catch (error) {
-            debugLogger.error("Twitter Retweet", "Failed to retweet", error);
-            throw error;
-          }
+        // Twitter steps
+        case "twitterCreateTweet":
+          result = await this.twitterHandler.executeCreateTweet(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "twitterSearchTweets": {
-          try {
-            const searchQuery = this.substituteVariables(step.searchQuery);
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Search Tweets", "Searching tweets", { searchQuery });
-
-            const params: Record<string, string> = {
-              query: searchQuery,
-              max_results: String(step.maxResults || 10),
-            };
-
-            if (step.startTime) {
-              params.start_time = new Date(step.startTime).toISOString();
-            }
-
-            if (step.endTime) {
-              params.end_time = new Date(step.endTime).toISOString();
-            }
-
-            const queryString = new URLSearchParams(params).toString();
-            const url = `https://api.twitter.com/2/tweets/search/recent?${queryString}`;
-
-            const oauth = this.generateTwitterOAuthHeader(
-              "GET",
-              url,
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.get(url, {
-              headers: { Authorization: oauth },
-            });
-
-            debugLogger.debug("Twitter Search Tweets", "Search completed", {
-              resultsCount: response.data.data?.length || 0,
-            });
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data.data || [];
-            }
-
-            result = response.data.data || [];
-          } catch (error) {
-            debugLogger.error("Twitter Search Tweets", "Failed to search tweets", error);
-            throw error;
-          }
+        case "twitterDeleteTweet":
+          result = await this.twitterHandler.executeDeleteTweet(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "twitterSendDM": {
-          try {
-            const userId = this.substituteVariables(step.userId).replace("@", "");
-            const text = this.substituteVariables(step.text);
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Send DM", "Sending direct message", { userId });
-
-            let recipientId = userId;
-            if (!/^\d+$/.test(userId)) {
-              const userOauth = this.generateTwitterOAuthHeader(
-                "GET",
-                `https://api.twitter.com/2/users/by/username/${userId}`,
-                apiKey,
-                apiSecret,
-                accessToken,
-                accessSecret
-              );
-
-              const userResponse = await axios.get(
-                `https://api.twitter.com/2/users/by/username/${userId}`,
-                {
-                  headers: { Authorization: userOauth },
-                }
-              );
-
-              recipientId = userResponse.data.data.id;
-            }
-
-            const body: Record<string, unknown> = { text };
-
-            if (step.mediaId) {
-              body.attachments = [{ media_id: this.substituteVariables(step.mediaId) }];
-            }
-
-            const dmOauth = this.generateTwitterOAuthHeader(
-              "POST",
-              `https://api.twitter.com/2/dm_conversations/with/${recipientId}/messages`,
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.post(
-              `https://api.twitter.com/2/dm_conversations/with/${recipientId}/messages`,
-              body,
-              {
-                headers: {
-                  Authorization: dmOauth,
-                  "Content-Type": "application/json",
-                },
-              }
-            );
-
-            debugLogger.debug("Twitter Send DM", "Direct message sent successfully");
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data;
-            }
-
-            result = response.data;
-          } catch (error) {
-            debugLogger.error("Twitter Send DM", "Failed to send direct message", error);
-            throw error;
-          }
+        case "twitterLikeTweet":
+          result = await this.twitterHandler.executeLikeTweet(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
           break;
-        }
 
-        case "twitterSearchUser": {
-          try {
-            const username = this.substituteVariables(step.username).replace("@", "");
-            const { apiKey, apiSecret, accessToken, accessSecret } =
-              await this.resolveTwitterCredentials(step);
-
-            debugLogger.debug("Twitter Search User", "Searching user", { username });
-
-            const oauth = this.generateTwitterOAuthHeader(
-              "GET",
-              `https://api.twitter.com/2/users/by/username/${username}`,
-              apiKey,
-              apiSecret,
-              accessToken,
-              accessSecret
-            );
-
-            const response = await axios.get(
-              `https://api.twitter.com/2/users/by/username/${username}`,
-              {
-                headers: { Authorization: oauth },
-              }
-            );
-
-            debugLogger.debug("Twitter Search User", "User found", {
-              userId: response.data.data?.id,
-            });
-
-            if (step.storeKey) {
-              this.variables[step.storeKey] = response.data.data;
-            }
-
-            result = response.data.data;
-          } catch (error) {
-            debugLogger.error("Twitter Search User", "Failed to search user", error);
-            throw error;
-          }
+        case "twitterRetweet":
+          result = await this.twitterHandler.executeRetweet(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
           break;
-        }
+
+        case "twitterSearchTweets":
+          result = await this.twitterHandler.executeSearchTweets(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
+          break;
+
+        case "twitterSendDM":
+          result = await this.twitterHandler.executeSendDM(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
+          break;
+
+        case "twitterSearchUser":
+          result = await this.twitterHandler.executeSearchUser(
+            step,
+            this.substituteVariables.bind(this),
+            this.resolveTwitterCredentials.bind(this),
+            variables
+          );
+          break;
+
+        // Variable modification steps
+        case "setVariable":
+          result = this.variableOperationHandler.executeSetVariable(
+            step,
+            this.substituteVariables.bind(this),
+            this.parseValue.bind(this),
+            variables
+          );
+          break;
+
+        case "modifyVariable":
+          result = this.variableOperationHandler.executeModifyVariable(
+            step,
+            this.substituteVariables.bind(this),
+            this.parseValue.bind(this),
+            variables
+          );
+          break;
+
+        default:
+          throw new Error(`Unknown step type: ${step.type}`);
       }
 
       const duration = performance.now() - startTime;
@@ -1086,111 +422,8 @@ export class AutomationExecutor {
   }
 
   /**
-   * Extract tweet ID from URL or return as-is if it's already an ID
-   */
-  private extractTweetId(input: string): string {
-    if (/^\d+$/.test(input)) {
-      return input;
-    }
-
-    try {
-      const url = new URL(input);
-      if (!/(twitter|x)\.com$/.test(url.hostname)) {
-        throw new Error("Invalid Twitter/X domain");
-      }
-      const parts = url.pathname.split("/");
-      if (parts.length >= 4 && parts[2] === "status" && /^\d+$/.test(parts[3])) {
-        return parts[3];
-      }
-    } catch {
-      // If parsing fails, return as-is
-    }
-
-    return input;
-  }
-
-  /**
-   * Generate Twitter OAuth 1.0a Authorization header
-   */
-  private generateTwitterOAuthHeader(
-    method: string,
-    url: string,
-    apiKey: string,
-    apiSecret: string,
-    accessToken: string,
-    accessSecret: string
-  ): string {
-    const timestamp = Math.floor(Date.now() / 1000).toString();
-    const nonce = crypto.randomBytes(32).toString("base64").replace(/\W/g, "");
-
-    const oauthParams: Record<string, string> = {
-      oauth_consumer_key: apiKey,
-      oauth_nonce: nonce,
-      oauth_signature_method: "HMAC-SHA1",
-      oauth_timestamp: timestamp,
-      oauth_token: accessToken,
-      oauth_version: "1.0",
-    };
-
-    // Parse URL parameters
-    const urlObj = new URL(url);
-    const urlParams: Record<string, string> = {};
-    urlObj.searchParams.forEach((value, key) => {
-      urlParams[key] = value;
-    });
-
-    // Combine all parameters
-    const allParams = { ...oauthParams, ...urlParams };
-
-    // Create parameter string
-    const sortedKeys = Object.keys(allParams).sort();
-    const paramString = sortedKeys
-      .map((key) => `${this.percentEncode(key)}=${this.percentEncode(allParams[key])}`)
-      .join("&");
-
-    // Create signature base string
-    const baseUrl = `${urlObj.protocol}//${urlObj.host}${urlObj.pathname}`;
-    const signatureBaseString = [
-      method.toUpperCase(),
-      this.percentEncode(baseUrl),
-      this.percentEncode(paramString),
-    ].join("&");
-
-    // Create signing key
-    const signingKey = `${this.percentEncode(apiSecret)}&${this.percentEncode(accessSecret)}`;
-
-    // Generate signature
-    const signature = crypto
-      .createHmac("sha1", signingKey)
-      .update(signatureBaseString)
-      .digest("base64");
-
-    oauthParams.oauth_signature = signature;
-
-    // Create Authorization header
-    const headerParts = Object.keys(oauthParams)
-      .sort()
-      .map((key) => `${this.percentEncode(key)}="${this.percentEncode(oauthParams[key])}"`)
-      .join(", ");
-
-    return `OAuth ${headerParts}`;
-  }
-
-  /**
-   * Percent-encode per OAuth spec (RFC 3986)
-   */
-  private percentEncode(str: string): string {
-    return encodeURIComponent(str)
-      .replace(/!/g, "%21")
-      .replace(/'/g, "%27")
-      .replace(/\(/g, "%28")
-      .replace(/\)/g, "%29")
-      .replace(/\*/g, "%2A");
-  }
-
-  /**
    * Evaluates a browser conditional node (elementExists or valueMatches)
-   * These conditions require a browser window and interact with the DOM
+   * Delegates to ConditionalEvaluator
    */
   async evaluateBrowserConditional(
     browserWindow: BrowserWindow | undefined,
@@ -1212,145 +445,18 @@ export class AutomationExecutor {
     conditionResult: boolean;
     effectiveSelector?: string | null;
   }> {
-    if (headless) {
-      if (!headlessExecutor) throw new Error("Headless executor not initialized");
-      return headlessExecutor.evaluateBrowserConditional({
-        browserConditionType: config.browserConditionType,
-        selector: this.substituteVariables(config.selector),
-        expectedValue: this.substituteVariables(config.expectedValue),
-        condition: config.condition,
-        transformType: config.transformType,
-        transformPattern: config.transformPattern,
-        transformReplace: config.transformReplace,
-        transformChars: config.transformChars,
-        parseAsNumber: config.parseAsNumber,
-      });
-    } else {
-      const wc = browserWindow.webContents;
-      const { browserConditionType, selector, expectedValue } = config;
-      if (!browserConditionType || !selector) {
-        throw new Error("browserConditionType and selector are required");
-      }
-      const startTime = performance.now();
-
-      debugLogger.debug("BrowserConditional", `Evaluating ${browserConditionType} condition`, {
-        selector,
-        expectedValue,
-      });
-
-      const runtimeSelector = this.substituteVariables(selector);
-
-      const applyTransform = (raw: string) => {
-        if (!raw) return raw;
-        const t = config.transformType || "none";
-        let s = raw;
-        if (t === "stripCurrency") {
-          s = s.replace(/[$€£,\s]/g, "");
-        } else if (t === "stripNonNumeric") {
-          s = s.replace(/[^0-9.-]/g, "");
-        } else if (t === "removeChars" && config.transformChars) {
-          const chars = config.transformChars.split("");
-          for (const c of chars) s = s.split(c).join("");
-        } else if (t === "regexReplace" && config.transformPattern) {
-          try {
-            const re = new RegExp(config.transformPattern, "g");
-            s = s.replace(re, config.transformReplace ?? "");
-          } catch (_e) {
-            debugLogger.warn("BrowserConditional", "Invalid regex pattern", {
-              pattern: config.transformPattern,
-            });
-          }
-        }
-        return s;
-      };
-
-      let conditionResult = false;
-
-      if (browserConditionType === "elementExists") {
-        conditionResult = await wc.executeJavaScript(`
-        (() => {
-          const sel = ${JSON.stringify(runtimeSelector)};
-          let el = null;
-          try { el = document.querySelector(sel); } catch(e) {}
-          if (!el) {
-            try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-          }
-          return !!el;
-        })();
-      `);
-        debugLogger.debug(
-          "BrowserConditional",
-          `Element ${conditionResult ? "found" : "not found"}`,
-          {
-            selector: runtimeSelector,
-          }
-        );
-      } else if (browserConditionType === "valueMatches") {
-        const rawValue: string = await wc.executeJavaScript(`
-        (() => {
-          const sel = ${JSON.stringify(runtimeSelector)};
-          let el = null;
-          try { el = document.querySelector(sel); } catch(e) {}
-          if (!el) {
-            try { const r = document.evaluate(sel, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null); el = r.singleNodeValue; } catch(e2) {}
-          }
-          return el?.innerText || "";
-        })();
-      `);
-        const transformed = applyTransform(rawValue);
-        const expected = this.substituteVariables(expectedValue || "");
-        const op = config.condition || "equals";
-
-        debugLogger.debug("BrowserConditional", "Value matching", {
-          rawValue,
-          transformed,
-          expected,
-          operator: op,
-        });
-
-        if (config.parseAsNumber) {
-          const a = parseFloat(transformed.replace(/[^0-9.-]/g, ""));
-          const b = parseFloat(expected.replace(/[^0-9.-]/g, ""));
-          conditionResult =
-            !isNaN(a) &&
-            !isNaN(b) &&
-            (op === "greaterThan"
-              ? a > b
-              : op === "lessThan"
-                ? a < b
-                : op === "contains"
-                  ? transformed.includes(expected)
-                  : a === b);
-          debugLogger.debug(
-            "BrowserConditional",
-            `Numeric comparison: ${a} ${op} ${b} = ${conditionResult}`
-          );
-        } else {
-          conditionResult =
-            op === "contains"
-              ? transformed.includes(expected)
-              : op === "greaterThan"
-                ? parseFloat(transformed) > parseFloat(expected)
-                : op === "lessThan"
-                  ? parseFloat(transformed) < parseFloat(expected)
-                  : transformed === expected;
-        }
-      }
-
-      const duration = performance.now() - startTime;
-      debugLogger.logOperation(
-        "BrowserConditional",
-        `Condition evaluated to: ${conditionResult}`,
-        duration
-      );
-
-      return { conditionResult, effectiveSelector: runtimeSelector };
-    }
+    return this.conditionalEvaluator.evaluateBrowserConditional(
+      browserWindow,
+      headless,
+      headlessExecutor,
+      config,
+      this.substituteVariables.bind(this)
+    );
   }
 
   /**
    * Evaluates a variable conditional node (variable-based conditions)
-   * These conditions do not require a browser window and work with stored variables
+   * Delegates to ConditionalEvaluator
    */
   evaluateVariableConditional(config: {
     variableConditionType?: string;
@@ -1360,329 +466,18 @@ export class AutomationExecutor {
   }): {
     conditionResult: boolean;
   } {
-    const { variableConditionType, variableName, expectedValue } = config;
-    if (!variableConditionType || !variableName) {
-      throw new Error("variableConditionType and variableName are required");
-    }
-    const startTime = performance.now();
-
-    debugLogger.debug("VariableConditional", `Evaluating ${variableConditionType} condition`, {
-      variableName,
-      expectedValue,
-    });
-
-    const resolvedVarName = this.substituteVariables(variableName);
-    const variableValue = this.getVariableValue(resolvedVarName);
-    let conditionResult = false;
-
-    if (variableConditionType === "variableExists") {
-      conditionResult =
-        variableValue !== undefined && variableValue !== null && variableValue !== "";
-      debugLogger.debug(
-        "VariableConditional",
-        `Variable ${conditionResult ? "exists" : "does not exist"}`,
-        {
-          variableName: resolvedVarName,
-          value: variableValue,
-        }
-      );
-    } else {
-      const expected = this.substituteVariables(expectedValue || "");
-      const actualStr =
-        variableValue === null || variableValue === undefined ? "" : String(variableValue);
-
-      if (config.parseAsNumber) {
-        const a = parseFloat(actualStr);
-        const b = parseFloat(expected);
-
-        if (isNaN(a) || isNaN(b)) {
-          conditionResult = false;
-          debugLogger.warn("VariableConditional", "Cannot parse as number", {
-            variableValue: actualStr,
-            expectedValue: expected,
-          });
-        } else {
-          if (variableConditionType === "variableEquals") {
-            conditionResult = a === b;
-          } else if (variableConditionType === "variableGreaterThan") {
-            conditionResult = a > b;
-          } else if (variableConditionType === "variableLessThan") {
-            conditionResult = a < b;
-          } else if (variableConditionType === "variableContains") {
-            // Contains doesn't make sense for numbers, fall back to string
-            conditionResult = actualStr.includes(expected);
-          }
-          debugLogger.debug(
-            "VariableConditional",
-            `Numeric comparison: ${a} ${variableConditionType} ${b} = ${conditionResult}`
-          );
-        }
-      } else {
-        // String comparison
-        if (variableConditionType === "variableEquals") {
-          conditionResult = actualStr === expected;
-        } else if (variableConditionType === "variableContains") {
-          conditionResult = actualStr.includes(expected);
-        } else if (variableConditionType === "variableGreaterThan") {
-          conditionResult = actualStr > expected;
-        } else if (variableConditionType === "variableLessThan") {
-          conditionResult = actualStr < expected;
-        }
-        debugLogger.debug(
-          "VariableConditional",
-          `String comparison: "${actualStr}" ${variableConditionType} "${expected}" = ${conditionResult}`
-        );
-      }
-    }
-
-    const duration = performance.now() - startTime;
-    debugLogger.logOperation(
-      "VariableConditional",
-      `Condition evaluated to: ${conditionResult}`,
-      duration
+    return this.conditionalEvaluator.evaluateVariableConditional(
+      config,
+      this.substituteVariables.bind(this),
+      this.getVariableValue.bind(this)
     );
-
-    return { conditionResult };
-  }
-
-  /**
-   * Execute OpenAI API call (GPT-4, GPT-3.5, etc.)
-   * Deterministic defaults: temperature 0, maxTokens 256, no streaming or tools.
-   */
-  private async executeAiOpenAI(step: StepAIOpenAI): Promise<string> {
-    const prompt = this.substituteVariables(step.prompt || "").trim();
-    const systemPrompt = this.substituteVariables(step.systemPrompt || "").trim();
-    const model = this.substituteVariables(step.model || "").trim();
-    const temperature = Math.max(0, Math.min(1, Number(step.temperature ?? 0)));
-    const maxTokens = Math.min(Math.max(1, Math.floor(Number(step.maxTokens ?? 256))), 4096);
-    const topPValue =
-      step.topP === undefined ? undefined : Math.max(0, Math.min(1, Number(step.topP)));
-    const timeoutMs = Math.min(Math.max(1000, Number(step.timeoutMs ?? 20000)), 120000);
-
-    if (!prompt) throw new Error("OpenAI step requires a prompt");
-    if (!model) throw new Error("OpenAI step requires a model");
-
-    const baseUrl = this.normalizeBaseUrl(
-      this.substituteVariables(step.baseUrl || ""),
-      "https://api.openai.com/v1"
-    );
-    const apiKey = await this.resolveOpenAIApiKey(step);
-    if (!apiKey) {
-      throw new Error("API key is required for OpenAI");
-    }
-
-    debugLogger.debug("AI OpenAI", "Preparing request", {
-      model,
-      baseUrl,
-      promptLength: prompt.length,
-      hasSystemPrompt: !!systemPrompt,
-      temperature,
-      maxTokens,
-      topP: topPValue,
-    });
-
-    const systemMessage = systemPrompt ? [{ role: "system", content: systemPrompt }] : [];
-    const userMessage = { role: "user", content: prompt };
-
-    const url = `${baseUrl}/chat/completions`;
-    const payload: Record<string, unknown> = {
-      model,
-      messages: [...systemMessage, userMessage],
-      temperature,
-      max_tokens: maxTokens,
-      n: 1,
-      stream: false,
-    };
-    if (topPValue !== undefined) payload.top_p = topPValue;
-
-    const response = await axios.post(url, payload, {
-      headers: {
-        Authorization: `Bearer ${apiKey}`,
-        "Content-Type": "application/json",
-      },
-      timeout: timeoutMs,
-    });
-
-    const content = response.data?.choices?.[0]?.message?.content;
-    if (!content) {
-      throw new Error("OpenAI returned an empty response");
-    }
-    return typeof content === "string" ? content.trim() : JSON.stringify(content);
-  }
-
-  /**
-   * Execute Anthropic (Claude) API call
-   * Deterministic defaults: temperature 0, maxTokens 256, no streaming or tools.
-   */
-  private async executeAiAnthropic(step: StepAIAnthropic): Promise<string> {
-    const prompt = this.substituteVariables(step.prompt || "").trim();
-    const systemPrompt = this.substituteVariables(step.systemPrompt || "").trim();
-    const model = this.substituteVariables(step.model || "").trim();
-    const temperature = Math.max(0, Math.min(1, Number(step.temperature ?? 0)));
-    const maxTokens = Math.min(Math.max(1, Math.floor(Number(step.maxTokens ?? 256))), 4096);
-    const topPValue =
-      step.topP === undefined ? undefined : Math.max(0, Math.min(1, Number(step.topP)));
-    const timeoutMs = Math.min(Math.max(1000, Number(step.timeoutMs ?? 20000)), 120000);
-
-    if (!prompt) throw new Error("Anthropic step requires a prompt");
-    if (!model) throw new Error("Anthropic step requires a model");
-
-    const baseUrl = this.normalizeBaseUrl(
-      this.substituteVariables(step.baseUrl || ""),
-      "https://api.anthropic.com"
-    );
-    const apiKey = await this.resolveAnthropicApiKey(step);
-    if (!apiKey) {
-      throw new Error("API key is required for Anthropic");
-    }
-
-    debugLogger.debug("AI Anthropic", "Preparing request", {
-      model,
-      baseUrl,
-      promptLength: prompt.length,
-      hasSystemPrompt: !!systemPrompt,
-      temperature,
-      maxTokens,
-      topP: topPValue,
-    });
-
-    const userMessage = { role: "user", content: prompt };
-
-    const url = `${baseUrl}/v1/messages`;
-    const payload: Record<string, unknown> = {
-      model,
-      max_tokens: maxTokens,
-      temperature,
-      messages: [userMessage],
-      stream: false,
-    };
-    if (systemPrompt) payload.system = systemPrompt;
-    if (topPValue !== undefined) payload.top_p = topPValue;
-
-    const response = await axios.post(url, payload, {
-      headers: {
-        "x-api-key": apiKey,
-        "anthropic-version": "2023-06-01",
-        "Content-Type": "application/json",
-      },
-      timeout: timeoutMs,
-    });
-
-    const content = response.data?.content?.[0]?.text || response.data?.content?.[0]?.content;
-    if (!content) {
-      throw new Error("Anthropic returned an empty response");
-    }
-    return String(content).trim();
-  }
-
-  /**
-   * Execute Ollama (local LLM) API call
-   * Deterministic defaults: temperature 0, maxTokens 256, no streaming or tools.
-   */
-  private async executeAiOllama(step: StepAIOllama): Promise<string> {
-    const prompt = this.substituteVariables(step.prompt || "").trim();
-    const systemPrompt = this.substituteVariables(step.systemPrompt || "").trim();
-    const model = this.substituteVariables(step.model || "").trim();
-    const temperature = Math.max(0, Math.min(1, Number(step.temperature ?? 0)));
-    const maxTokens = Math.min(Math.max(1, Math.floor(Number(step.maxTokens ?? 256))), 4096);
-    const topPValue =
-      step.topP === undefined ? undefined : Math.max(0, Math.min(1, Number(step.topP)));
-    const timeoutMs = Math.min(Math.max(1000, Number(step.timeoutMs ?? 20000)), 120000);
-
-    if (!prompt) throw new Error("Ollama step requires a prompt");
-    if (!model) throw new Error("Ollama step requires a model");
-
-    const baseUrl = this.normalizeBaseUrl(
-      this.substituteVariables(step.baseUrl || ""),
-      "http://localhost:11434"
-    );
-
-    debugLogger.debug("AI Ollama", "Preparing request", {
-      model,
-      baseUrl,
-      promptLength: prompt.length,
-      hasSystemPrompt: !!systemPrompt,
-      temperature,
-      maxTokens,
-      topP: topPValue,
-    });
-
-    const systemMessage = systemPrompt ? [{ role: "system", content: systemPrompt }] : [];
-    const userMessage = { role: "user", content: prompt };
-
-    const url = `${baseUrl}/api/chat`;
-    const payload: Record<string, unknown> = {
-      model,
-      messages: [...systemMessage, userMessage],
-      stream: false,
-      options: {
-        temperature,
-        top_p: topPValue,
-        num_predict: maxTokens,
-      },
-    };
-
-    const response = await axios.post(url, payload, {
-      timeout: timeoutMs,
-      headers: { "Content-Type": "application/json" },
-    });
-
-    const content = response.data?.message?.content || response.data?.response;
-    if (!content) {
-      throw new Error("Ollama returned an empty response");
-    }
-    return String(content).trim();
-  }
-
-  private normalizeBaseUrl(baseUrl: string, fallback: string): string {
-    const value = baseUrl?.trim() || fallback;
-    return value.replace(/\/+$/, "");
-  }
-
-  private async resolveOpenAIApiKey(step: StepAIOpenAI): Promise<string | null> {
-    if (step.credentialId) {
-      const credential = await getCredential(step.credentialId);
-      if (!credential) throw new Error("OpenAI credential not found");
-      const fromStore =
-        credential.data.apiKey ||
-        credential.data.key ||
-        credential.data.token ||
-        credential.data.accessToken;
-      if (fromStore) return this.substituteVariables(fromStore);
-      throw new Error("OpenAI credential is missing an API key value");
-    }
-
-    if (step.apiKey) {
-      return this.substituteVariables(step.apiKey);
-    }
-
-    return null;
-  }
-
-  private async resolveAnthropicApiKey(step: StepAIAnthropic): Promise<string | null> {
-    if (step.credentialId) {
-      const credential = await getCredential(step.credentialId);
-      if (!credential) throw new Error("Anthropic credential not found");
-      const fromStore =
-        credential.data.apiKey ||
-        credential.data.key ||
-        credential.data.token ||
-        credential.data.accessToken;
-      if (fromStore) return this.substituteVariables(fromStore);
-      throw new Error("Anthropic credential is missing an API key value");
-    }
-
-    if (step.apiKey) {
-      return this.substituteVariables(step.apiKey);
-    }
-
-    return null;
   }
 
   /**
    * Resolve Twitter credentials from either credentialId or direct fields
+   * Delegates to TwitterStepHandler
    */
-  private async resolveTwitterCredentials(step: {
+  private resolveTwitterCredentials(step: {
     credentialId?: string;
     apiKey?: string;
     apiSecret?: string;
@@ -1694,71 +489,17 @@ export class AutomationExecutor {
     accessToken: string;
     accessSecret: string;
   }> {
-    // If credentialId is provided, fetch from store
-    if (step.credentialId) {
-      const credential = await getCredential(step.credentialId);
-      if (!credential || credential.type !== "twitter") {
-        throw new Error("Invalid or missing Twitter credential");
-      }
-      return {
-        apiKey: credential.data.apiKey || "",
-        apiSecret: credential.data.apiSecret || "",
-        accessToken: credential.data.accessToken || "",
-        accessSecret: credential.data.accessSecret || "",
-      };
-    }
-
-    // Otherwise use direct fields (legacy support)
-    return {
-      apiKey: this.substituteVariables(step.apiKey || ""),
-      apiSecret: this.substituteVariables(step.apiSecret || ""),
-      accessToken: this.substituteVariables(step.accessToken || ""),
-      accessSecret: this.substituteVariables(step.accessSecret || ""),
-    };
+    return TwitterStepHandler.resolveTwitterCredentials(step);
   }
 
   /**
    * Resolve Discord credentials from store or direct fields
+   * Delegates to DiscordStepHandler
    */
-  private async resolveDiscordCredentials(step: {
+  private resolveDiscordCredentials(step: {
     credentialId?: string;
     botToken?: string;
   }): Promise<{ botToken: string }> {
-    if (step.credentialId) {
-      const credential = await getCredential(step.credentialId);
-      if (!credential || credential.type !== "discord") {
-        throw new Error("Invalid or missing Discord credential");
-      }
-      return {
-        botToken: credential.data.botToken || "",
-      };
-    }
-
-    return {
-      botToken: this.substituteVariables(step.botToken || ""),
-    };
-  }
-
-  /**
-   * Lightweight Discord API client helper
-   */
-  private async discordRequest(
-    method: "GET" | "POST" | "DELETE" | "PUT",
-    path: string,
-    data: unknown,
-    botToken: string
-  ) {
-    const baseUrl = "https://discord.com/api/v10";
-    const url = `${baseUrl}${path}`;
-
-    return axios({
-      method,
-      url,
-      data,
-      headers: {
-        Authorization: `Bot ${botToken}`,
-        "Content-Type": "application/json",
-      },
-    });
+    return DiscordStepHandler.resolveDiscordCredentials(step);
   }
 }
