+++ src/main/handlers/index.ts
@@ -1,0 +1,10 @@
+// Export all handlers from a single entry point
+
+export { AiStepHandler } from "./aiStepHandler";
+export { ApiCallHandler } from "./apiCallHandler";
+export { BrowserStepHandler } from "./browserStepHandler";
+export { ConditionalEvaluator } from "./conditionalEvaluator";
+export { DiscordStepHandler } from "./discordStepHandler";
+export { TwitterStepHandler } from "./twitterStepHandler";
+export { VariableManager } from "./variableManager";
+export { VariableOperationHandler } from "./variableOperationHandler";
